Reverted the old mechanism: runs

+++ src/gpu/blas1/ExSUM.FPE.EX.8.cl
@@ -308,7 +308,7 @@ void ExSUM(
 ////////////////////////////////////////////////////////////////////////////////
 __kernel __attribute__((reqd_work_group_size(MERGE_WORKGROUP_SIZE, 1, 1)))
 void ExSUMComplete(
-    __global double *d_Superacc,
+    __global long *d_Superacc,
     __global long *d_PartialSuperaccs,
     uint PartialSuperaccusCount
 ) {
